Fix backgrounds that had too many colors
grit doesn't do a great job at reducing palette sizes, so this has been
done manually. All the images were added to GIMP as different layers, then
it was asked to generate a 256 color palette. Then, mtPaint was used to
fix the transparent color to be (255, 0, 255) exactly and to save it as a
PNG file.